refactor(fsmeta): move raftstore runtime wiring into exec
